more fixes

+++ solanaceae/ngc_ft1/cubic.hpp
@@ -35,7 +35,7 @@ struct CUBIC : public FlowOnly {
 		// TODO: api for how much data we should send
 		// take time since last sent into account
 		// respect max_byterate_allowed
-		size_t canSend(void) override;
+		int64_t canSend(void) override;
 
 		// get the list of timed out seq_ids
 		//std::vector<SeqIDType> getTimeouts(void) const override;
